+++ notebridge/hello_page.py
@@ -1,0 +1,29 @@
+hello_page_html = '''
+<html>
+    <head>
+        <style>
+            body {
+                display: flex;
+                flex-direction: column;
+                align-items: flex-start;
+                justify-content: flex-start;
+                padding: 26px;
+            }
+            h1 {
+                font-size: 44px;
+            }
+            p {
+                font-family: -apple-system, BlinkMacSystemFont,
+                             "Segoe UI", Roboto, "Helvetica Neue", Arial,
+                             sans-serif, "Apple Color Emoji", "Segoe UI Emoji",
+                             "Segoe UI Symbol";
+                font-size: 24px;            
+            }
+        </style>
+    </head>
+    <body>
+        <h1>🚀</h1>
+        <p>Haha! You need to send a POST request to this endpoint to interact with agent. — Notebridge</p>
+    </body>
+</html>
+'''